+++ clients/admin-ui/src/features/privacy-experience/PrivacyExperienceForm.tsx
@@ -29,15 +29,21 @@ import {
 import { getSelectedRegionIds } from "~/features/privacy-experience/form/helpers";
 import { selectAllLanguages } from "~/features/privacy-experience/language.slice";
 import {
-  selectAllPrivacyNotices,
   selectPage as selectNoticePage,
   selectPageSize as selectNoticePageSize,
   useGetAllPrivacyNoticesQuery,
 } from "~/features/privacy-notices/privacy-notices.slice";
 import {
+  selectAllProperties,
+  selectPage as selectPropertyPage,
+  selectPageSize as selectPropertyPageSize,
+  useGetAllPropertiesQuery,
+} from "~/features/properties/property.slice";
+import {
   ComponentType,
   ExperienceConfigCreate,
   ExperienceTranslation,
+  LimitedPrivacyNoticeResponseSchema,
   SupportedLanguage,
 } from "~/types/api";
 
@@ -74,8 +80,10 @@ export const PrivacyExperienceConfigColumnLayout = ({
 );
 
 export const PrivacyExperienceForm = ({
+  allPrivacyNotices,
   onSelectTranslation,
 }: {
+  allPrivacyNotices: LimitedPrivacyNoticeResponseSchema[];
   onSelectTranslation: (t: ExperienceTranslation) => void;
 }) => {
   const router = useRouter();
@@ -86,7 +94,6 @@ export const PrivacyExperienceForm = ({
   const noticePage = useAppSelector(selectNoticePage);
   const noticePageSize = useAppSelector(selectNoticePageSize);
   useGetAllPrivacyNoticesQuery({ page: noticePage, size: noticePageSize });
-  const allPrivacyNotices = useAppSelector(selectAllPrivacyNotices);
 
   const getPrivacyNoticeName = (id: string) => {
     const notice = allPrivacyNotices.find((n) => n.id === id);
@@ -108,6 +115,11 @@ export const PrivacyExperienceForm = ({
     const name = language ? language.name : t.language;
     return `${name}${t.is_default ? " (Default)" : ""}`;
   };
+
+  const propertyPage = useAppSelector(selectPropertyPage);
+  const propertyPageSize = useAppSelector(selectPropertyPageSize);
+  useGetAllPropertiesQuery({ page: propertyPage, size: propertyPageSize });
+  const allProperties = useAppSelector(selectAllProperties);
 
   if (editingStyle) {
     return (
@@ -153,45 +165,67 @@ export const PrivacyExperienceForm = ({
         isRequired
         variant="stacked"
       />
-      <CustomSelect
-        name="component"
-        id="component"
-        options={componentTypeOptions}
-        label="Experience Type"
-        variant="stacked"
-        isDisabled={!!values.component}
-        isRequired
-      />
-      <Collapse
-        in={
-          values.component && values.component !== ComponentType.PRIVACY_CENTER
-        }
-        animateOpacity
-      >
-        <Box p="1px">
-          <CustomSwitch
-            name="dismissable"
-            id="dismissable"
-            label="Modal is dismissable"
+      {values.component !== ComponentType.TCF_OVERLAY ? (
+        <>
+          <CustomSelect
+            name="component"
+            id="component"
+            options={componentTypeOptions}
+            label="Experience Type"
             variant="stacked"
+            isDisabled={!!values.component}
+            isRequired
           />
-        </Box>
-      </Collapse>
-      <Divider />
-      <Heading fontSize="md" fontWeight="semibold">
-        Privacy notices
-      </Heading>
+          <Collapse
+            in={
+              values.component &&
+              values.component !== ComponentType.PRIVACY_CENTER
+            }
+            animateOpacity
+          >
+            <Box p="1px">
+              <CustomSwitch
+                name="dismissable"
+                id="dismissable"
+                label="Modal is dismissable"
+                variant="stacked"
+              />
+            </Box>
+          </Collapse>
+        </>
+      ) : null}
       <ScrollableList
-        addButtonLabel="Add privacy notice"
-        allItems={allPrivacyNotices.map((n) => n.id)}
-        values={values.privacy_notice_ids ?? []}
-        setValues={(newValues) =>
-          setFieldValue("privacy_notice_ids", newValues)
-        }
-        getItemLabel={getPrivacyNoticeName}
+        label="Associated properties"
+        addButtonLabel="Add property"
+        idField="id"
+        nameField="name"
+        allItems={allProperties.map((property) => ({
+          id: property.id,
+          name: property.name,
+        }))}
+        values={values.properties ?? []}
+        setValues={(newValues) => setFieldValue("properties", newValues)}
         draggable
       />
       <Divider />
+      {values.component !== ComponentType.TCF_OVERLAY ? (
+        <>
+          <Heading fontSize="md" fontWeight="semibold">
+            Privacy notices
+          </Heading>
+          <ScrollableList
+            addButtonLabel="Add privacy notice"
+            allItems={allPrivacyNotices.map((n) => n.id)}
+            values={values.privacy_notice_ids ?? []}
+            setValues={(newValues) =>
+              setFieldValue("privacy_notice_ids", newValues)
+            }
+            getItemLabel={getPrivacyNoticeName}
+            draggable
+          />
+          <Divider />
+        </>
+      ) : null}
       <Text as="h2" fontWeight="600">
         Locations & Languages
       </Text>