+++ clients/admin-ui/src/features/privacy-notices/privacy-notices.slice.ts
@@ -103,6 +103,7 @@ export const {
   usePatchPrivacyNoticesMutation,
   useLimitedPatchPrivacyNoticesMutation,
   useGetPrivacyNoticeByIdQuery,
+  useLazyGetPrivacyNoticeByIdQuery,
   usePostPrivacyNoticeMutation,
 } = privacyNoticesApi;