+++ clients/admin-ui/src/types/api/models/ExperienceConfigCreate.ts
@@ -5,6 +5,7 @@
 import type { ComponentType } from "./ComponentType";
 import type { ExperienceTranslationCreate } from "./ExperienceTranslationCreate";
 import type { PrivacyNoticeRegion } from "./PrivacyNoticeRegion";
+import type { Property } from "./Property";
 
 /**
  * Schema for creating Experience Configs via the API
@@ -21,4 +22,5 @@ export type ExperienceConfigCreate = {
   component: ComponentType;
   privacy_notice_ids?: Array<string>;
   translations?: Array<ExperienceTranslationCreate>;
+  properties?: Array<Property>;
 };