+++ clients/fides-js/src/components/ConsentModal.tsx
@@ -37,7 +37,11 @@ const ConsentModal = ({
           <div />
           <CloseButton
             ariaLabel="Close modal"
-            onClick={closeButton.onClick}
+            onClick={
+              window.Fides.options.fidesPreviewMode
+                ? () => {}
+                : closeButton.onClick
+            }
             hidden={window.Fides.options.preventDismissal}
           />
         </div>