+++ src/fides/api/schemas/property.py
@@ -1,12 +1,17 @@
-from __future__ import annotations
-
-import re
 from enum import Enum
-from typing import Any, Dict, Optional
-
-from pydantic import root_validator
+from typing import List, Optional
 
 from fides.api.schemas.base_class import FidesSchema
+
+
+class MinimalPrivacyExperience(FidesSchema):
+    """
+    Minimal representation of a privacy experience, contains enough information
+    to select experiences by name in the UI and an ID to link the selections in the database.
+    """
+
+    id: str
+    name: str
 
 
 class PropertyType(Enum):
@@ -14,18 +19,13 @@ class PropertyType(Enum):
     other = "Other"
 
 
+class MinimalProperty(FidesSchema):
+    id: str
+    name: str
+
+
 class Property(FidesSchema):
     name: str
     type: PropertyType
-    key: Optional[str] = None
-
-    @root_validator
-    def generate_key(cls, values: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Generate the property key from the name if not supplied
-        1) remove invalid characters
-        2) replace spaces with underscores
-        """
-        name = re.sub(r"[^a-zA-Z0-9._<> -]", "", values["name"].lower().strip())
-        values["key"] = re.sub(r"\s+", "_", name)
-        return values
+    id: Optional[str] = None
+    experiences: List[MinimalPrivacyExperience]